Added OrderFill Record. TODO: Fix Portfolio Tests

+++ src/main/java/com/trading/domain/ValidatedOrder.java
@@ -6,7 +6,7 @@ public record ValidatedOrder(
         String strategyId,
         String symbol,
         Side side,
-        long quantity,
+        double quantity,
         double price,
         long timestamp
 ) implements Order {}
